set temperature units to double, added slope calculation and display on frontend

+++ src/main/java/nl/timnederhoff/tools/maischperfect/model/highcharts/Point.java
@@ -2,9 +2,9 @@ package nl.timnederhoff.tools.maischperfect.model.highcharts;
 
 public class Point {
 	long x;
-	int y;
+	double y;
 
-	public Point(long x, int y) {
+	public Point(long x, double y) {
 		this.x = x;
 		this.y = y;
 	}
@@ -13,7 +13,7 @@ public class Point {
 		return x;
 	}
 
-	public int getY() {
+	public double getY() {
 		return y;
 	}
 }
